+++ ving/generator/nuxtpages.mjs
@@ -25,6 +25,14 @@ const columns = (schema) => {
                 </template>
             </Column>`;
         }
+        else if (prop.name == 'userId') {
+            out += `
+        <Column field="props.userId" header="User Id" sortable>
+            <template #body="slotProps">
+                <UserProfileLink :user="slotProps.data.related?.user" />
+            </template>
+        </Column>`;
+        }
         else if (prop.type == 'date') {
             out += `
             <Column field="props.${prop.name}" header="${makeLabel(prop.name)}" sortable>
@@ -173,6 +181,11 @@ const viewProps = (schema) => {
             else if (['boolean', 'enum'].includes(prop.type)) {
                 out += `
             <div><b>${makeLabel(prop.name)}</b>: {{enum2label(${schema.kind.toLowerCase()}.props?.${prop.name}, ${schema.kind.toLowerCase()}.options?.${prop.name})}}</div>
+            `;
+            }
+            else if (prop.name == 'userId') {
+                out += `
+            <div><b>${makeLabel(prop.name)}</b>: <UserProfileLink :user="${schema.kind.toLowerCase()}.related?.user" /></div>
             `;
             }
             else if (prop.type != 'virtual') {